requirements.txt
transformers 
langchain 
pandas>=2.0.0
openpyxl>=3.1.0
streamlit>=1.22.0
openai>=1.0.0
python-dotenv>=1.0.0
torch 
accelerate
